Cleanup of the tables for WOFF1 fonts
