+++ c2pa-font-handler/src/woff1/directory_test.rs
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-//! Tests for woff1 directory module
+//! Tests for WOFF1 directory module
 
 use std::io::Cursor;
